+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- support WebUSB fallback
+
 ## [0.0.50](https://github.com/andy23512/tangent-cc-lib/releases/tag/0.0.50)
 
 - add actionAndPhraseHash property to chord object